umap_service/requirements.txt
fastapi
uvicorn
numpy
umap-learn
scikit-learn
